add component FormForMap
